% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medio.R
\name{regla_punto_medio}
\alias{regla_punto_medio}
\title{Integracion mediante la Regla del Punto Medio}
\usage{
regla_punto_medio(a, b, f)
}
\arguments{
\item{a}{extremo inferior del intervalo}

\item{b}{extremo superior del intervalo}

\item{f}{funcion que vamos a integrar}
}
\value{
La aproximacion de la integral en el intervalo
}
\description{
Esta funcion nos calcula una aproximacion de la integral de f en el intervalo
de a hasta b.
}
\details{
En nuestra entrada f, tenemos que definir nuestra funcion. Es decir
f es del tipo funcion.
}
\examples{
regla_punto_medio(0,1,function(x){return(x)})
regla_punto_medio(2,2.5,function(x){return(x+2)})

}
