% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectangulo.R
\name{regla_rectangulo}
\alias{regla_rectangulo}
\title{Integracion mediante la Regla del Rectangulo}
\usage{
regla_rectangulo(a, b, f)
}
\arguments{
\item{a}{extremo inferior del intervalo}

\item{b}{extremo superior del intervalo}

\item{f}{funcion que vamos a integrar}
}
\value{
La aproximacion de la integral en el intervalo
}
\description{
Esta funcion nos calcula una aproximacion de la integral de f en el intervalo
de a hasta b, mediante la Regla del Rectangulo.
}
\details{
En nuestra entrada f, tenemos que definir nuestra funcion. Es decir
f es del tipo funcion.
}
\examples{
regla_rectangulo(1.5,3,function(x){return(x^2)})
regla_rectangulo(2,3,function(x){return(sin(x))})

}
\seealso{
\link{regla_punto_medio}
}
