% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpson.R
\name{regla_simpson}
\alias{regla_simpson}
\title{Integracion mediante la Regla de Simpson}
\usage{
regla_simpson(a, b, f)
}
\arguments{
\item{a}{extremo inferior del intervalo}

\item{b}{extremo superior del intervalo}

\item{f}{funcion que vamos a integrar}
}
\value{
La aproximacion de la integral en el intervalo
}
\description{
Esta funcion nos calcula una aproximacion de la integral de f en el intervalo
de a hasta b, mediante la Regla de Simpson.
}
\details{
En nuestra entrada f, tenemos que definir nuestra funcion. Es decir
f es del tipo funcion.
}
\examples{
regla_simpson(1,3,function(x){return(x^2+x^3+x)})
regla_simpson(2,4,function(x){return(cos(x)+x)})

}
