% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trapecio.R
\name{regla_trapecio}
\alias{regla_trapecio}
\title{Integracion mediante la Regla del Trapecio}
\usage{
regla_trapecio(a, b, f)
}
\arguments{
\item{a}{extremo inferior del intervalo}

\item{b}{extremo superior del intervalo}

\item{f}{funcion que vamos a integrar}
}
\value{
La aproximacion de la integral en el intervalo
}
\description{
Esta funcion nos calcula una aproximacion de la integral de f en el intervalo
de a hasta b, mediante la Regla del Trapecio.
}
\details{
En nuestra entrada f, tenemos que definir nuestra funcion. Es decir
f es del tipo funcion.
}
\examples{
regla_trapecio(1.5,3,function(x){return(x^2+x^3)})
regla_trapecio(2.5,3.14,function(x){return(cos(x))})

}
